Split into multiple Dockerfiles with docker-compose

+++ check_docker.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-docker run --pull always --rm -i \
-  -v $(pwd)/hadolint.yaml:/.config/hadolint.yaml \
-  hadolint/hadolint < docker/Dockerfile
+for dir in *; do
+  if [ -d "${dir}" ]; then
+    echo "${dir}"
+    docker run --pull always --rm -i \
+      -v $(pwd)/hadolint.yaml:/.config/hadolint.yaml \
+      hadolint/hadolint < ${dir}/Dockerfile
+  fi
+done
